Revise tests
Gcode unknown commands and command errors are now test exceptions
`ASSERT TEST="{printer.whatever} is False"` in tests allows template
 evaluated runtime assertions

fix a couple broken tests that did not error because of the above
