+++ pyproject.toml
@@ -50,6 +50,5 @@ dev = [
   ]
 
 [tool.pytest.ini_options]
-testpaths = [
-  'test',
-]
+pythonpath = ["."]
+testpaths = ["test"]
