+++ test/klippy/z_tilt_ng.test
@@ -16,8 +16,6 @@ G1 Z2.924972
 ACCEPT
 G1 Z2.959972
 ACCEPT
-G1 Z2.924972
-ACCEPT
 
 # Report position
 G1 Z5 X10 Y10 F6000
